using exact copy of FastNoise2.cs (with partial keyword) to simplify maintenance
