+++ UnityProject/Assets/FastNoise2/FastNoise2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cab6ba7e169093e4387c65b36fe89fac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
